+++ mapbox-maps-ios/Tests/MapboxMapsTests/Integration Tests/TileCoverIntegrationTests.swift
@@ -1,0 +1,23 @@
+import XCTest
+@_spi(Experimental) import MapboxMaps
+
+class TileCoverIntegrationTests: MapViewIntegrationTestCase {
+    func testTileCover() {
+        mapView.mapboxMap.setCamera(to: CameraOptions(zoom: 5.5))
+        let tileIds = mapView.mapboxMap.tileCover(for: TileCoverOptions(tileSize: 512, minZoom: 0, maxZoom: 22, roundZoom: false))
+        XCTAssertFalse(tileIds.isEmpty)
+        for tileId in tileIds {
+            XCTAssertEqual(tileId.z, 5) // sanity check
+        }
+    }
+
+    func testTileCoverDefaultParameters() throws {
+        throw XCTSkip("Default value resolution will be fixed in MAPSNAT-1024")
+        mapView.mapboxMap.setCamera(to: CameraOptions(zoom: 5.5))
+        let tileIds = mapView.mapboxMap.tileCover(for: TileCoverOptions())
+        XCTAssertFalse(tileIds.isEmpty)
+        for tileId in tileIds {
+            XCTAssertEqual(tileId.z, 5) // sanity check
+        }
+    }
+}